Copious fussing about and still no skybox
I have no idea what is happening. For some reason, the window is always
black, even though the camera should be encapsulated in a skybox. The VBO
is valid, I'm fairly certain the draw calls are valid, the textures are
valid, the shaders compile without any complaints, and no matter how hard
I try it simply will not work.